+++ include/sdf/SDFImpl.hh
@@ -110,13 +110,13 @@ namespace sdf
     /// \param[in] _path Path on disk.
     public: void SetFilePath(const std::string &_path);
 
-    /// \brief Set the spec version that this was parsed from.
+    /// \brief Set the spec version that this was originally parsed from.
     /// \param[in] _version Spec version string.
-    public: void SetParsedVersion(const std::string &_version);
+    public: void SetOriginalVersion(const std::string &_version);
 
-    /// \brief Get the spec version that this was parsed from.
+    /// \brief Get the spec version that this was originally parsed from.
     /// \return Spec version string.
-    public: const std::string &ParsedVersion() const;
+    public: const std::string &OriginalVersion() const;
 
     /// \brief Get the version
     /// \return The version as a string