+++ src/SDF_TEST.cc
@@ -548,13 +548,13 @@ TEST(SDF, FilePath)
 }
 
 /////////////////////////////////////////////////
-TEST(SDF, ParsedVersion)
+TEST(SDF, OriginalVersion)
 {
   sdf::SDF sdf;
-  EXPECT_TRUE(sdf.ParsedVersion().empty());
+  EXPECT_TRUE(sdf.OriginalVersion().empty());
 
-  sdf.SetParsedVersion("1.5");
-  EXPECT_EQ("1.5", sdf.ParsedVersion());
+  sdf.SetOriginalVersion("1.5");
+  EXPECT_EQ("1.5", sdf.OriginalVersion());
 }
 
 /////////////////////////////////////////////////