+++ src/parser_TEST.cc
@@ -77,8 +77,8 @@ TEST(Parser, NameUniqueness)
     EXPECT_FALSE(sdf::recursiveSameTypeUniqueNames(sdf->Root()));
     EXPECT_EQ(path, sdf->FilePath());
     EXPECT_EQ(path, sdf->Root()->FilePath());
-    EXPECT_EQ("1.6", sdf->ParsedVersion());
-    EXPECT_EQ("1.6", sdf->Root()->ParsedVersion());
+    EXPECT_EQ("1.6", sdf->OriginalVersion());
+    EXPECT_EQ("1.6", sdf->Root()->OriginalVersion());
   }
 
   // Check an SDF file with sibling elements of different types (model, light)
@@ -90,8 +90,8 @@ TEST(Parser, NameUniqueness)
     EXPECT_FALSE(sdf::recursiveSiblingUniqueNames(sdf->Root()));
     EXPECT_EQ(path, sdf->FilePath());
     EXPECT_EQ(path, sdf->Root()->FilePath());
-    EXPECT_EQ("1.6", sdf->ParsedVersion());
-    EXPECT_EQ("1.6", sdf->Root()->ParsedVersion());
+    EXPECT_EQ("1.6", sdf->OriginalVersion());
+    EXPECT_EQ("1.6", sdf->Root()->OriginalVersion());
   }
 
   // Check an SDF file with sibling elements of the same type (link)
@@ -103,8 +103,8 @@ TEST(Parser, NameUniqueness)
     EXPECT_FALSE(sdf::recursiveSameTypeUniqueNames(sdf->Root()));
     EXPECT_EQ(path, sdf->FilePath());
     EXPECT_EQ(path, sdf->Root()->FilePath());
-    EXPECT_EQ("1.6", sdf->ParsedVersion());
-    EXPECT_EQ("1.6", sdf->Root()->ParsedVersion());
+    EXPECT_EQ("1.6", sdf->OriginalVersion());
+    EXPECT_EQ("1.6", sdf->Root()->OriginalVersion());
   }
 
   // Check an SDF file with sibling elements of the same type (joint)
@@ -116,8 +116,8 @@ TEST(Parser, NameUniqueness)
     EXPECT_FALSE(sdf::recursiveSameTypeUniqueNames(sdf->Root()));
     EXPECT_EQ(path, sdf->FilePath());
     EXPECT_EQ(path, sdf->Root()->FilePath());
-    EXPECT_EQ("1.6", sdf->ParsedVersion());
-    EXPECT_EQ("1.6", sdf->Root()->ParsedVersion());
+    EXPECT_EQ("1.6", sdf->OriginalVersion());
+    EXPECT_EQ("1.6", sdf->Root()->OriginalVersion());
   }
 
   // Check an SDF file with sibling elements of different types (link, joint)
@@ -129,8 +129,8 @@ TEST(Parser, NameUniqueness)
     EXPECT_FALSE(sdf::recursiveSiblingUniqueNames(sdf->Root()));
     EXPECT_EQ(path, sdf->FilePath());
     EXPECT_EQ(path, sdf->Root()->FilePath());
-    EXPECT_EQ("1.6", sdf->ParsedVersion());
-    EXPECT_EQ("1.6", sdf->Root()->ParsedVersion());
+    EXPECT_EQ("1.6", sdf->OriginalVersion());
+    EXPECT_EQ("1.6", sdf->Root()->OriginalVersion());
   }
 
   // Check an SDF file with sibling elements of the same type (collision)
@@ -142,8 +142,8 @@ TEST(Parser, NameUniqueness)
     EXPECT_FALSE(sdf::recursiveSameTypeUniqueNames(sdf->Root()));
     EXPECT_EQ(path, sdf->FilePath());
     EXPECT_EQ(path, sdf->Root()->FilePath());
-    EXPECT_EQ("1.6", sdf->ParsedVersion());
-    EXPECT_EQ("1.6", sdf->Root()->ParsedVersion());
+    EXPECT_EQ("1.6", sdf->OriginalVersion());
+    EXPECT_EQ("1.6", sdf->Root()->OriginalVersion());
   }
 
   // Check an SDF file with sibling elements of the same type (visual)
@@ -155,8 +155,8 @@ TEST(Parser, NameUniqueness)
     EXPECT_FALSE(sdf::recursiveSiblingUniqueNames(sdf->Root()));
     EXPECT_EQ(path, sdf->FilePath());
     EXPECT_EQ(path, sdf->Root()->FilePath());
-    EXPECT_EQ("1.6", sdf->ParsedVersion());
-    EXPECT_EQ("1.6", sdf->Root()->ParsedVersion());
+    EXPECT_EQ("1.6", sdf->OriginalVersion());
+    EXPECT_EQ("1.6", sdf->Root()->OriginalVersion());
   }
 
   // Check an SDF file with cousin elements of the same type (collision)
@@ -169,8 +169,8 @@ TEST(Parser, NameUniqueness)
     EXPECT_TRUE(sdf::recursiveSiblingUniqueNames(sdf->Root()));
     EXPECT_EQ(path, sdf->FilePath());
     EXPECT_EQ(path, sdf->Root()->FilePath());
-    EXPECT_EQ("1.6", sdf->ParsedVersion());
-    EXPECT_EQ("1.6", sdf->Root()->ParsedVersion());
+    EXPECT_EQ("1.6", sdf->OriginalVersion());
+    EXPECT_EQ("1.6", sdf->Root()->OriginalVersion());
   }
 
   // Check an SDF file with cousin elements of the same type (visual)
@@ -183,8 +183,8 @@ TEST(Parser, NameUniqueness)
     EXPECT_TRUE(sdf::recursiveSiblingUniqueNames(sdf->Root()));
     EXPECT_EQ(path, sdf->FilePath());
     EXPECT_EQ(path, sdf->Root()->FilePath());
-    EXPECT_EQ("1.6", sdf->ParsedVersion());
-    EXPECT_EQ("1.6", sdf->Root()->ParsedVersion());
+    EXPECT_EQ("1.6", sdf->OriginalVersion());
+    EXPECT_EQ("1.6", sdf->Root()->OriginalVersion());
   }
 }