+++ Changelog.md
@@ -22,7 +22,7 @@
 1. Converter: fix memory leak pointed out by ASan.
     * [Pull request 638](https://bitbucket.org/osrf/sdformat/pull-requests/638)
 
-1. Access the original parsed version of an SDF document with `Element::ParsedVersion`.
+1. Access the original parsed version of an SDF document with `Element::OriginalVersion`.
     * [Pull request 640](https://bitbucket.org/osrf/sdformat/pull-requests/640)
 
 ### SDFormat 9.0.0 (2019-12-10)